move InstrumentedHook to metrics package
